Electrica_backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
cssselect2==0.7.0
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
fonttools==4.38.0
html5lib==1.1
idna==3.4
lxml==4.9.2
oscrypto==1.3.0
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
pydyf==0.5.0
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyPDF3==1.0.6
pyphen==0.13.2
python-bidi==0.4.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.13
weasyprint==57.2
webencodings==0.5.1
zopfli==0.2.2